payment method

+++ cart_panel.py
@@ -3,6 +3,7 @@ from cart_button import CartButton
 from discount_button import DiscountButton
 from panel import Panel
 from pay_button import PayButton
+from payment_method import PaymentMethod
 from utils import *
 
 SPACE = 7
@@ -11,22 +12,39 @@ SPACE = 7
 class CartPanel(Panel):
     def __init__(self, galaxy, rect, border_width, border_radius):
         super().__init__(galaxy, rect, border_width, border_radius)
-        self.discount_percentage = 0
         self.discount_button = None
+        self.discount_percentage = 0
+        self.payment_method_index = 0
+        self.payment_methods = [
+                'Cartão de débito',
+                'Cartão de crédito',
+                'Pix',
+                'Dinheiro'
+            ]
+        self.payment_method_button = None
         self.pay_button = None
 
     def update(self, time_passed, event_list):
         super().update(time_passed, event_list)
 
-        # destroy discount button
-        if self.discount_button:
-            self.discount_button.kill()  # remove from galaxy
-
-        # destroy pay button
-        if self.pay_button:
-            self.pay_button.kill()  # remove from galaxy
+        # destroy payment method, discount and pay buttons
+        if self.discount_button: self.discount_button.kill()  
+        if self.payment_method_button: self.payment_method_button.kill()
+        if self.pay_button: self.pay_button.kill()  
 
         if self.has_cart_items() > 0:
+            # prepare payment method button text
+            payment_method_button_text = self.payment_methods[self.payment_method_index]
+
+            # create payment method button and add to galaxy
+            self.payment_method_button = PaymentMethod(
+                self.galaxy, payment_method_button_text, CENTER, self.payment_method_pressed_event)
+            self.payment_method_button.rect = pygame.Rect(
+                self.last_cart_item().rect)
+            self.payment_method_button.rect.y += self.payment_method_button.rect.height + 3*SPACE
+            self.payment_method_button.rect.height /= 1.4
+            self.galaxy.add_entity(self.payment_method_button)
+
             # prepare discount button text
             if self.discount_percentage == 0:
                 discount_button_text = 'Sem desconto'
@@ -38,9 +56,8 @@ class CartPanel(Panel):
             # create discount button and add to galaxy
             self.discount_button = DiscountButton(
                 self.galaxy, discount_button_text, CENTER, self.discount_pressed_event)
-            self.discount_button.rect = pygame.Rect(
-                self.last_cart_item().rect)
-            self.discount_button.rect.y += self.discount_button.rect.height + 3*SPACE
+            self.discount_button.rect = pygame.Rect(self.payment_method_button.rect)
+            self.discount_button.rect.y += self.discount_button.rect.height + SPACE
             self.galaxy.add_entity(self.discount_button)
 
             # prepare pay button text
@@ -54,8 +71,8 @@ class CartPanel(Panel):
                 self.galaxy, pay_button_text, CENTER, None)
             self.pay_button.rect = pygame.Rect(
                 self.last_cart_item().rect)
-            self.pay_button.rect.bottom = self.rect.bottom - 4*SPACE
-            self.pay_button.rect.height *= 1.3
+            self.pay_button.rect.bottom = self.rect.bottom - 2*SPACE
+            #self.pay_button.rect.height *= 1.3
             self.galaxy.add_entity(self.pay_button)
 
     def render(self, surface):
@@ -65,6 +82,12 @@ class CartPanel(Panel):
         if len(self.cart_items_list()) > 0:
             return True
         return False
+
+    def payment_method_pressed_event(self, button):
+        if self.payment_method_index == 3:
+            self.payment_method_index = 0
+        else:
+            self.payment_method_index += 1
 
     def discount_pressed_event(self, button):
         if self.discount_percentage == 15:
